+++ ving/generator/nuxtpages.mjs
@@ -350,6 +350,7 @@ const ${name.toLowerCase()} = useVingRecord({
     },
 });
 await ${name.toLowerCase()}.fetch()
+onBeforeRouteLeave(() => ${name.toLowerCase()}.dispose());
 
 const breadcrumbs = [
     { label: '${makeWords(name)}s', to: '/${name.toLowerCase()}' },