+++ docs/change-log.md
@@ -15,6 +15,7 @@ outline: deep
 * Fixed: documentation for Related Records is broken #86
 * Implemented: write docs for utils #89
 * Fixed: drizzle kit is now interactive #93
+* Fixed: cannot dispose useVingRecord on edit page #102
 
 ## 2024-04-10
 * Implemented: filterQualifier: true should be in the examples for all relation ids #96